pyproject.toml
[project]
name = "ocr-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chandra-ocr",
    "google-cloud-documentai>=3.7.0",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.4",
]

[tool.uv.sources]
chandra-ocr = { git = "https://github.com/datalab-to/chandra" }
